fix: health bar updates and player ID handling
- Change HealthChangeEvent entity_id from u32 to u64 for full player IDs
- Add MyClientId resource to store client ID at startup
- Add mark_local_player system to properly identify local player
- Update health bar to use stored client ID for reliable updates
- Add comprehensive logging for debugging health sync
- Comment out position sync performance logs to reduce noise
- Fix all ID comparisons to handle u64 consistently

Fixes health bar not updating when player takes damage in multiplayer

+++ shared/src/protocol.rs
@@ -357,8 +357,8 @@ impl MapEntities for ProjectileDespawnEvent {
 /// Event sent when entity health changes (for efficient health updates)
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Reflect)]
 pub struct HealthChangeEvent {
-    /// Network entity ID
-    pub entity_id: u32,
+    /// Network entity ID - u64 to support full player IDs
+    pub entity_id: u64,
     /// New health value
     pub new_health: f32,
     /// Maximum health value
